ai-chat-app/requirements.txt
Flask
dotenv
google-generativeai
requests
